requirements.txt
antiorm==1.2.1
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==8.0.1
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
db==0.1.1
dbus-python==1.2.16
defer==1.0.6
deluge==2.0.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
entrypoints==0.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Blogging==1.2.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-FileUpload==0.5.0
Flask-Login==0.5.0
Flask-MySQLdb==1.0.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
GeoIP==1.3.2
googlemaps==4.6.0
greenlet==1.1.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
more-itertools==4.2.0
mysqlclient==2.1.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
outcome==1.1.0
pandas==1.4.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pygame==1.9.6
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-libtorrent==1.1.13
python-slugify==5.0.2
pytz==2021.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
rencode==1.0.6
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.3
service-identity==18.1.0
setproctitle==1.1.10
shortuuid==1.0.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.2
SQLAlchemy==1.4.17
ssh-import-id==5.10
systemd-python==234
text-unidecode==1.3
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
visitor==0.1.3
wadllib==1.3.3
Werkzeug==2.0.1
wsproto==1.1.0
WTForms==2.3.3
youtube-dl==2020.3.24
zipp==1.0.0
zope.interface==4.7.1
